ClothingDTO implementation, clothing-inventory relation fixed, new store relation added

+++ src/main/java/com/macedonia/macedonia/services/impl/StoreServiceImpl.java
@@ -1,10 +1,13 @@
 package com.macedonia.macedonia.services.impl;
 
+import com.macedonia.macedonia.entities.Inventory;
 import com.macedonia.macedonia.entities.Store;
+import com.macedonia.macedonia.persistence.InventoryDAO;
 import com.macedonia.macedonia.persistence.StoreDAO;
 import com.macedonia.macedonia.services.StoreService;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -12,12 +15,32 @@ import java.util.Optional;
 public class StoreServiceImpl implements StoreService {
 
     private final StoreDAO storeDAO;
+    private final InventoryDAO inventoryDAO;
 
-    public StoreServiceImpl(StoreDAO storeDAO){
+    public StoreServiceImpl(StoreDAO storeDAO, InventoryDAO inventoryDAO){
+
         this.storeDAO = storeDAO;
+        this.inventoryDAO = inventoryDAO;
     }
+    /*@Override
+    public Store createStore(Store store) {
+        return storeDAO.save(store);
+    }*/
+
     @Override
     public Store createStore(Store store) {
+        if (store.getInventory() == null) {
+            Inventory inventory = new Inventory();
+            inventory.setName("Inventario de " + store.getName());
+            inventory.setLocation(store.getLocation());
+            inventory.setTotalItems(0);
+            inventory.setLastUpdated(LocalDateTime.now());
+            store.setInventory(inventory);
+        } else {
+            Inventory existingInventory = inventoryDAO.findById(store.getInventory().getId())
+                    .orElseThrow(() -> new RuntimeException("Inventario no encontrada con ID: " + store.getInventory().getId()));
+            store.setInventory(existingInventory);
+        }
         return storeDAO.save(store);
     }
 
@@ -30,19 +53,23 @@ public class StoreServiceImpl implements StoreService {
             updatedStore.setLocation(store.getLocation());
             return storeDAO.save(updatedStore);
         } else {
-            throw new RuntimeException("Store not found with ID: " + id);
+            throw new RuntimeException("Tienda no encontrada con ID: " + id);
         }
     }
 
     @Override
     public void deleteStore(Long id) {
-        storeDAO.deleteById(id);
+        Optional<Store> store = storeDAO.findById(id); // Busca el Store por ID
+        if (store.isEmpty()) { // Verifica si el Store no existe
+            throw new RuntimeException("Tienda no encontrada con ID: " + id);
+        }
+        storeDAO.deleteById(id); // Si existe, procede a eliminar
     }
 
     @Override
     public Store getStoreById(Long id) {
         return storeDAO.findById(id).orElseThrow(()->
-                new RuntimeException(("Store not found with ID: " + id)));
+                new RuntimeException(("Tienda no encontrada con ID: " + id)));
     }
 
     @Override
